refactored presenter

+++ pages/api/presenters/DeletePresenter.ts
@@ -9,11 +9,12 @@ export default async function handler(
   res: NextApiResponse
 ) {
   if (req.method === "DELETE") {
-    // const session = await getServerSession(req, res, authOptions);
-    // if (!session) {
-    //   return res.status(401).json({ message: "Please signin to delete." });
-    // }
+    const session = await getServerSession(req, res, authOptions);
+    if (!session) {
+      return res.status(401).json({ message: "Please signin to delete." });
+    }
     const presenterId = req.body;
+    console.log("req.body preseterId", req.body);
 
     try {
       const result = await prisma.gamePresenter.delete({
